More smooth curves
Implemented the path.quadTo function instead of lineTo which made the
curves more smooth.
It's not entirely smooth yet but it's a bit better.
If it's not enough it might work to increase the amount of points that's
registered. However that might affect performance.

Commented out the lineTo function for comparing.