Created hand animation
Created a new animation gif file. Had to import the WebView class to
include the gif in the app.
Removed the old image. And created an assets folder to store the gif in.
Changed the color of the view that the gif resides in because the color
code was strange and didn't work in photo shop.
Removed the description strings since they are no longer needed.

This is not scalable to other devices. I had to fit the WebView view
precisely to the same size as the gif or there would've been white
spaces or the gif would not be entirely visible. Also red online that
this method doesn't work on some devices.