Updated point in the end of list
Added a completely different point with the same value as the first one
in the end of the list containing the points. Changed the name of the
list in case there was some kind of conflict between the list in the
test class and the list in the non-test class.